got rid of network menu controller & network menu res
