+++ src/ca/bcit/jfx/res/MainWindow.fxml
@@ -51,7 +51,9 @@
 					<content>
 						<VBox alignment="TOP_CENTER" spacing="8.0">
 							<children>
-								<fx:include fx:id="networkMenu" source="/ca/bcit/jfx/res/NetworkMenu.fxml" />
+								<!--<fx:include fx:id="networkMenu" source="/ca/bcit/jfx/res/NetworkMenu.fxml" />-->
+								<Button onAction="#onNew" text="New" GridPane.rowIndex="1" />
+								<Button onAction="#onSave" text="Save" GridPane.rowIndex="2" />
 								<Button onAction="#onLoad" text="Load" GridPane.columnIndex="1" />
 								<RadioButton fx:id="RBNoneChose" onAction="#noneChose" selected="true" text="None" toggleGroup="$RBGroup">
 									<toggleGroup>
